+++ src/argilla/tasks/rich.py
@@ -12,11 +12,21 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from .database import app as database_app
-from .datasets import app as datasets_app
-from .login import app as login_app
-from .logout import app as logout_app
-from .server import app as server_app
-from .training import app as training_app
-from .users import app as users_app
-from .workspaces import app as workspaces_app
+from typing import TYPE_CHECKING, Any
+
+from rich.panel import Panel
+from rich.table import Table
+
+if TYPE_CHECKING:
+    from rich.console import RenderableType
+
+# TODO: update colors after consulting it with UI expert
+_ARGILLA_BORDER_STYLE = "red"
+
+
+def get_argilla_themed_table(title: str, **kwargs: Any) -> Table:
+    return Table(title=title, border_style=_ARGILLA_BORDER_STYLE, **kwargs)
+
+
+def get_argilla_themed_panel(renderable: "RenderableType", **kwargs: Any) -> Panel:
+    return Panel(renderable=renderable, border_style=_ARGILLA_BORDER_STYLE, **kwargs)